Refactor and modularise playground as IDE
