+++ src/components/IDE/__tests__/Output.tsx
@@ -1,7 +1,7 @@
 import { shallow } from 'enzyme'
 import * as React from 'react'
 
-import { ResultOutput } from '../../reducers/states'
+import { ResultOutput } from '../../../reducers/states'
 import Output from '../Output'
 
 test('Output renders correctly', () => {
